Fixed Contact Design Issue
When people submitted a message, the looping animation appeared to be a
loading screen. So, they would wait for an exorbitant amout of time to
get some sort of success message. I created a new animation to show a
check, and end the animation.

This is a good example of user testing revealing a flaw.