Downgrading Bundler to allow Netlify Continuous Deployment
Netlify doesn't run bundler 2.0 or above. In order to deploy using 
Netlify with Bundler 2.0, I neet to use the Netlify CLI. The problem 
with using the Netlify CLI is that NetlifyCMS no longer works, which was 
one of the main goals of this site update.

So, downgrading bundler to the latest version Netlify uses is the best 
solution for now.