First Round of Dark Mode Switch Functionality
I laid the groundwork for allowing the darkmode switch to affect the 
site by adding a "global" class to the body tag.

I also enabled some functionality, but need to put in the ability to 
swap dark/light GIFs and fix other bugs that come with such a 
fundamental design change.