Fully Functioning Dark/Light Mode Switch
The dark mode switch is now running and successfully swapping out a 
bunch of classes on the website to make it appear darker/lighter.

There is one issue, though. The menu button doesn't load in the 
light/dark gif on a switch flip. It only does it on a menu button click. 
So I need to figure out a solution for that, perhaps triggering a menu 
button click on switch flip.