Potential Fix for Issue #1, Temporarily Fixed Issue #4
I made better use of InfinitScroll events to run the Isotope sorting 
method after new objects are appeneded. Hopefully this fixes the jumbled 
loading on iPad.

As for the loading animation, I found a workaround. I replaced the "view 
more" button with a loading animation. This isn't the most intuitive, 
and I need to come up with a graphic that better telegraphs that 
something is loading while also being a clickable button. But I also 
enabled prefill on InfinitScroll, so you should never need to click view 
more, which means this might be a non-issue now.