Long overdue update
I fixed a few broken social links cause of changes to my Instagram and deletion of my YouTube channel.

I also fixed the annoying shadow clipping under the navigation pane issue.

I added a 404 error page.

I also added an award to one of the projects.

Basically, the site is now prepped for going back to the CMS and uploading a bunch of stuff!